Move clang-specific tests inside the clang plugin.
I want to be able to see what's a core test and what's a test for a specific plug, now that there's a difference. The triggering event is that I'm about to add some more clang integrations tests so I can then make them pass, and I don't want to make the tangle worse to do it.
