+++ .gitignore
@@ -2,11 +2,11 @@
 *.o
 *.so
 /dxr/static/js/templates.js
-/tests/*/target
-/tests/*/temp
-/tests/*/code/code
-/tests/*/code2/code
-/tests/*/dxr.config
+**/tests/*/target
+**/tests/*/temp
+**/tests/*/code/code
+**/tests/*/code2/code
+**/tests/*/dxr.config
 /vagrantconfig_local.yaml
 /.vagrant
 /build
