+++ dxr/plugins/clang/tests/test_anon_ns/code/main2.cpp
