+++ dxr/plugins/clang/tests/test_c_vardecl/code/extern.h
