+++ dxr/plugins/clang/tests/test_c_vardecl/dxr.config.in
