+++ dxr/plugins/clang/tests/test_clang.py
