+++ dxr/plugins/clang/tests/test_functions.py
