+++ dxr/plugins/clang/tests/test_namespaces.py
