+++ dxr/plugins/clang/tests/test_type_search.py
